Moved buy amount selections to dedicated component
